Refactor submit button and use on Dashboard
