+++ src/angular/planit/src/app/dashboard/dashboard.component.html
@@ -45,7 +45,9 @@
             <a class="button button-small"
                (click)="planService.export()"
                tooltip="Download your plan as a CSV">Download</a>
-            <button type="button" class="button button-small">Submit</button>
+            <app-submit-plan-button class="button button-small">
+              Submit
+            </app-submit-plan-button>
           </footer>
         </div>
       </div>
