+++ src/angular/planit/src/app/dashboard/review-plan/review-plan.component.ts
@@ -1,15 +1,11 @@
 import { Component, OnInit, ViewChild } from '@angular/core';
 
-import { ToastrService } from 'ngx-toastr';
 import { Subscription } from 'rxjs/Rx';
 
 import { PlanService } from '../../core/services/plan.service';
 import { RiskService } from '../../core/services/risk.service';
 import { UserService } from '../../core/services/user.service';
 import { OrgRiskRelativeOption, Organization, Risk, relativeOptionToNumber } from '../../shared';
-import {
-  ConfirmationModalComponent
-} from '../../shared/confirmation-modal/confirmation-modal.component';
 import { AccordionReviewState } from './tabs/adaptation-accordion-state.service';
 
 import { environment } from '../../../environments/environment';
@@ -27,8 +23,6 @@ enum ReviewPlanTabs {
 })
 export class ReviewPlanComponent implements OnInit {
 
-  @ViewChild('confirmSubmitModal') confirmSubmitModal: ConfirmationModalComponent;
-
   public activeTab = ReviewPlanTabs.OVERVIEW;
   public organization: Organization;
   public risks: Map<string, Risk[]>;
@@ -39,40 +33,12 @@ export class ReviewPlanComponent implements OnInit {
   constructor(private accordionState: AccordionReviewState,
               public planService: PlanService,
               private riskService: RiskService,
-              private toastr: ToastrService,
               private userService: UserService) { }
 
   ngOnInit() {
     this.userService.current().subscribe(user => this.organization = user.primary_organization);
     this.riskService.groupByWeatherEvent().subscribe(riskMap => this.sortAndSetRisks(riskMap));
     this.accordionState.isOpen = {};
-  }
-
-  public submitPlan() {
-    this.confirmSubmitModal.confirm({
-      title: 'Submit your plan',
-      tagline: 'Are you sure you\'re ready to submit your adaptation plan?',
-      confirmButtonClass: 'button-primary',
-      confirmText: 'Submit'
-    }).onErrorResumeNext().switchMap(() => {
-      return this.planService.submit();
-    }).subscribe(() => {
-      const success = `
-        Thanks for submitting your adaptation plan!
-        You will receive a copy of your submission via email.
-      `;
-      this.toastr.success(success);
-    }, () => {
-      const error = `
-        There was an unexpected issue submitting your plan.
-        Please try again in a few hours.
-        If you continue to have trouble, contact support@temperate.io to submit your plan.
-      `;
-      this.toastr.error(error, undefined, {
-        timeOut: 10000,
-        closeButton: true
-      });
-    });
   }
 
   private sortAndSetRisks(riskMap: Map<string, Risk[]>) {
