+++ src/angular/planit/src/app/shared/submit-plan-button/submit-plan-button.component.html
@@ -1,0 +1,3 @@
+<div (click)="submitPlan()"><ng-content></ng-content></div>
+
+<app-confirmation-modal #confirmSubmitModal></app-confirmation-modal>
